+++ changelog.md
@@ -7,6 +7,8 @@
 - Fix issue with setting reset
 - Try to work around ultra-rare bug where output passed as undefined (kwin bug?)
 - Improve stack based layout keyboard shortcuts
+- Add KWin layout
+- Add window dragging settings
 
 ### 1.2.0